remove toolbarutil, because it is only used in one place
