+++ app/src/main/java/ch/admin/bag/dp3t/stats/DiagramView.java
@@ -16,7 +16,6 @@ import android.graphics.Paint;
 import android.graphics.Path;
 import android.graphics.Rect;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
 import androidx.annotation.Nullable;
